+++ lib/flint_dart.dart
@@ -6,6 +6,7 @@ library;
 export 'security.dart'; // or specific ones like:
 export 'src/app.dart';
 export 'src/context.dart';
+export 'src/controller.dart';
 export 'src/routing/router.dart';
 export 'src/middleware/middleware.dart';
 export 'src/env_parser.dart';