+++ lib/src/cli/make_resource_command.dart
@@ -6,7 +6,7 @@ import 'package:flint_dart/src/cli/commands.dart';
 class MakeResourceCommand extends FlintCommand {
   MakeResourceCommand()
       : super(
-            'make:resource', '  Creates controller, routes and registers them');
+            '--make-resource', '  Creates controller, routes and registers them');
 
   @override
   Future<void> execute(List<String> args) async {